fe changes v1

+++ Aurora/backend/src/repositories/serviceRepository.js
@@ -7,7 +7,7 @@ class ServiceRepository {
       SELECT 
         id, 
         name, 
-        price, 
+        price::float, 
         duration_minutes AS "durationMinutes" 
       FROM services 
       ORDER BY name ASC;
